+++ writeopia_ui/src/commonMain/kotlin/io/writeopia/ui/drawer/content/CodeBlockDrawer.kt
@@ -55,14 +55,14 @@ private const val CODE_BLOCK_LINE_NUMBER_KEY = "codeBlockLineNumber"
 class CodeBlockDrawer(
     private val dragIconWidth: Dp = 16.dp,
     private val config: DrawConfig,
-    private val onSelected: (Boolean, Int) -> Unit = { _, _ -> },
-    private val onDragHover: (Int) -> Unit,
+    private val onSelected: (Boolean, Double) -> Unit = { _, _ -> },
+    private val onDragHover: (Double) -> Unit,
     private val onDragStart: () -> Unit = {},
     private val onDragStop: () -> Unit = {},
     private val moveRequest: (Action.Move) -> Unit = {},
     private val enabled: Boolean,
     private val isDesktop: Boolean,
-    private val receiveExternalFile: (List<ExternalFile>, Int) -> Unit,
+    private val receiveExternalFile: (List<ExternalFile>, Double) -> Unit,
     private val messageDrawer: @Composable RowScope.() -> SimpleTextDrawer,
 ) : StoryStepDrawer {
 
@@ -199,7 +199,7 @@ class CodeBlockDrawer(
         }
     }
 
-    private fun handleDrag(position: Int, data: DropInfo?) {
+    private fun handleDrag(position: Double, data: DropInfo?) {
         onDragHover(position)
 
         if (data != null) {
